sensors and processor integrated into application

+++ proj/MSH-backend/src/main/java/mysweethome/MSHbackend/Models/SensorData.java
@@ -14,11 +14,11 @@ public class SensorData {
     @Id
     private int data_source_id;
     @NotBlank
-    private int timestamp;
+    private Long timestamp;
     @NotBlank
     private String sensor_information;
 
-    public SensorData(int data_source_id, @NotBlank int timestamp, @NotBlank String sensor_information) {
+    public SensorData(int data_source_id, @NotBlank Long timestamp, @NotBlank String sensor_information) {
         this.data_source_id = data_source_id;
         this.timestamp = timestamp;
         this.sensor_information = sensor_information;
@@ -30,10 +30,10 @@ public class SensorData {
     public void setData_source_id(int data_source_id) {
         this.data_source_id = data_source_id;
     }
-    public int getTimestamp() {
+    public Long getTimestamp() {
         return timestamp;
     }
-    public void setTimestamp(int timestamp) {
+    public void setTimestamp(Long timestamp) {
         this.timestamp = timestamp;
     }
     public String getSensor_information() {
